+++ src/test/java/org/cojen/tupl/LimitCapacityTest.java
@@ -33,9 +33,9 @@ public class LimitCapacityTest {
 
     @Before
     public void createTempDb() throws Exception {
-        DatabaseConfig config = new DatabaseConfig().checkpointSizeThreshold(0);
-        config.directPageAccess(false);
-        mDb = (LocalDatabase) newTempDatabase(config);
+        mConfig = new DatabaseConfig().checkpointSizeThreshold(0);
+        mConfig.directPageAccess(false);
+        mDb = newTempDatabase(mConfig);
     }
 
     @After
@@ -44,7 +44,8 @@ public class LimitCapacityTest {
         mDb = null;
     }
 
-    protected LocalDatabase mDb;
+    protected DatabaseConfig mConfig;
+    protected Database mDb;
 
     @Test
     public void basic() throws Exception {
@@ -76,7 +77,7 @@ public class LimitCapacityTest {
         assertTrue(size < mDb.capacityLimit());
     }
 
-    private static void trim(LocalDatabase db, Index ix) throws Exception {
+    private static void trim(Database db, Index ix) throws Exception {
         db.capacityLimitOverride(-1);
         try {
             Cursor c = ix.newCursor(Transaction.BOGUS);
@@ -96,5 +97,107 @@ public class LimitCapacityTest {
         } finally {
             db.capacityLimitOverride(0);
         }
+    }
+
+    @Test
+    public void fragmentMess() throws Exception {
+        // Test with direct pointers.
+        fragmentMess(1_000_000, 100);
+    }
+
+    @Test
+    public void fragmentMessInline() throws Exception {
+        // Test with inline content.
+        fragmentMess(100_000, 10);
+    }
+
+    @Test
+    public void fragmentMessIndirect() throws Exception {
+        // Test with indirect pointers.
+        fragmentMess(4_000_000, 400);
+    }
+
+    private void fragmentMess(int size, int minFreed) throws Exception {
+        mDb.capacityLimit(size);
+        Index ix = mDb.openIndex("test");
+
+        // Allocate the root node.
+        ix.store(null, "hello".getBytes(), "world".getBytes());
+
+        Database.Stats stats = mDb.stats();
+        long total = stats.totalPages();
+
+        // Value is too large.
+        try {
+            byte[] value = new byte[size];
+            ix.store(null, "key".getBytes(), value);
+            fail();
+        } catch (DatabaseFullException e) {
+            // Expected.
+        }
+
+        stats = mDb.stats();
+        long delta = stats.totalPages() - total;
+        assertTrue(delta >= minFreed);
+        assertEquals(delta, stats.freePages());
+
+        assertEquals(null, ix.load(null, "key".getBytes()));
+
+        // Smaller value should work.
+        byte[] value = new byte[size / 2];
+        ix.store(null, "key".getBytes(), value);
+    }
+
+    @Test
+    public void fragmentRecovery() throws Exception {
+        // Tests that large value is in the redo log, even though it's not in the database.
+        // The opposite behavior is worse -- in the database but not in the redo log. This can
+        // happen if writing to the redo log fails. This is why redo is written first.
+
+        mDb.capacityLimit(1_000_000L);
+        Index ix = mDb.openIndex("test");
+
+        byte[] value = randomStr(new java.util.Random(), 1_000_000);
+
+        try {
+            ix.store(null, "key".getBytes(), value);
+            fail();
+        } catch (DatabaseFullException e) {
+            // Expected.
+        }
+
+        // Reopen without capacity limit.
+        mDb = reopenTempDatabase(mDb, mConfig);
+        ix = mDb.openIndex("test");
+
+        // Value was re-inserted.
+        fastAssertArrayEquals(value, ix.load(null, "key".getBytes()));
+    }
+
+    @Test
+    public void fragmentRollback() throws Exception {
+        // Tests that large value is fully rolled back when an explicit transaction is used.
+
+        mDb.capacityLimit(1_000_000L);
+        Index ix = mDb.openIndex("test");
+
+        byte[] value = randomStr(new java.util.Random(), 1_000_000);
+
+        Transaction txn = mDb.newTransaction();
+        try {
+            ix.store(txn, "key".getBytes(), value);
+            fail();
+        } catch (DatabaseFullException e) {
+            // Expected.
+        } finally {
+            txn.reset();
+        }
+
+        // Reopen without capacity limit.
+        mDb = reopenTempDatabase(mDb, mConfig);
+        ix = mDb.openIndex("test");
+
+        // Value is still gone.
+        assertEquals(null, ix.load(null, "key".getBytes()));
     }
 }